requirements.txt
openai==1.66.3
langgraph==0.3.11
langchain-openai==0.3.8
pandas
pydantic==2.10.6
python-dotenv==1.0.1
pytest
jupyter
ipykernel
tavily-python
langchain-core==0.3.45
langchain==0.3.21
langchain-community==0.3.20
duckduckgo-search
wikipedia